feature - 채팅방 송신 DTO 수정
### 한 것
- 메세지 전송시 필요한 DTO 패키지 수정
