+++ core/src/main/java/com/lastone/core/dto/message/ChatMessageReqDto.java
@@ -1,5 +1,6 @@
 package com.lastone.core.dto.message;
 
+import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -9,4 +10,13 @@ public class ChatMessageReqDto {
     private Long senderId;
     private Long receiverId;
     private String content;
+
+    private ChatMessageReqDto() {}
+
+    @Builder
+    public ChatMessageReqDto(Long senderId, Long receiverId, String content) {
+        this.senderId = senderId;
+        this.receiverId = receiverId;
+        this.content = content;
+    }
 }
